解决点击通知后其他通知数据发生改变的bug---》修改requestCode；修改审批页确定提示框样式

+++ app/src/main/java/com/definesys/dmportal/main/adapter/MsgRecycleViewAdapter.java
@@ -125,7 +125,7 @@ public class MsgRecycleViewAdapter extends RecyclerView.Adapter<RecyclerView.Vie
                                 ARouter.getInstance()
                                         .build(ARouterConstants.ApprovalLeaveInfoActivity)
                                         .withString("leaveId", myMessage.getMessageExtend())
-                                        .withInt("type", myMessage.getMessageExtend2())
+                                        .withInt("type", myMessage.getMessageExtend2().intValue())
                                         .withObject("date", myMessage.getSendTime())
                                         .withString("approvalContent", myMessage.getMessageContent())
                                         .navigation();
@@ -144,7 +144,7 @@ public class MsgRecycleViewAdapter extends RecyclerView.Adapter<RecyclerView.Vie
                                 ARouter.getInstance()
                                         .build(ARouterConstants.ApprovalApplyInfoActivity)
                                         .withString("applyId", myMessage.getMessageExtend())
-                                        .withInt("type", myMessage.getMessageExtend2())
+                                        .withInt("type", myMessage.getMessageExtend2().intValue())
                                         .withString("content",myMessage.getMessageContent())
                                         .withObject("date",myMessage.getSendTime())
                                         .navigation();
